feat: add separate ui component CookiesToastContainer
